service/intel-core/requirements.txt
aiokafka==0.13.0
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1
billiard==4.2.4
black==26.3.1
boto3==1.42.87
botocore==1.42.87
celery==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.7
distro==1.9.0
fastapi==0.135.3
google-auth==2.49.2
google-genai==1.72.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
isort==8.0.1
jmespath==1.1.0
joblib==1.5.3
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
numpy==2.4.4
orjson==3.11.8
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.6
pluggy==1.6.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
pytokens==0.4.1
PyYAML==6.0.3
redis==5.2.1
requests==2.33.1
s3transfer==0.16.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==0.52.1
tenacity==9.1.4
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
